requirements.txt
alembic==1.7.5
aniso8601==9.0.1
anyio==3.3.4
apispec==5.1.1
asgiref==3.4.1
click==8.0.3
dynaconf==3.1.7
fastapi==0.70.0
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
motor==2.5.1
pydantic==1.8.2
PyJWT==1.7.1
pymongo==3.12.1
PyMySQL==1.0.2
pytz==2021.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
SQLAlchemy-serializer==1.4.1
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
Werkzeug==2.0.2
